+++ packages/squiggle-lang/src/reducer/lambda.ts
@@ -2,8 +2,8 @@ import { LocationRange } from "peggy";
 import { Expression } from "../expression/index.js";
 import { Value } from "../value/index.js";
 import { Bindings } from "./bindings.js";
-import * as Context from "./Context.js";
-import { ReducerContext } from "./Context.js";
+import * as Context from "./context.js";
+import { ReducerContext } from "./context.js";
 import { ErrorMessage, REArityError } from "./ErrorMessage.js";
 import * as SqError from "./IError.js";
 import { ReducerFn } from "./index.js";
