--- packages/squiggle-lang/src/utility/E_O.ts
@@ -1,6 +1,0 @@
-export const fmap = <A, B>(
-  x: A | undefined,
-  fn: (v: A) => B
-): B | undefined => {
-  return x === undefined ? undefined : fn(x);
-};
